[ADD] 75% of events checkers and updaters done.

+++ SDL2-2.0.8/include/SDL_config_minimal.h
@@ -40,8 +40,8 @@
 typedef unsigned int size_t;
 typedef signed char int8_t;
 typedef unsigned char uint8_t;
-typedef signed short int16_t;
-typedef unsigned short uint16_t;
+typedef signed int int16_t;
+typedef unsigned int uint16_t;
 typedef signed int int32_t;
 typedef unsigned int uint32_t;
 typedef signed long long int64_t;
