+++ SDL2_mixer-2.0.4/external/libmodplug-0.8.9.0/src/load_dmf.cpp
@@ -492,7 +492,7 @@ dmfexit:
 
 typedef struct DMF_HNODE
 {
-	short int left, right;
+	int int left, right;
 	BYTE value;
 } DMF_HNODE;
