+++ SDL2_mixer-2.0.4/external/mpg123-1.25.6/src/metaprint.c
@@ -318,7 +318,7 @@ void print_id3_tag(mpg123_handle *mh, int long_id3, FILE *out)
 	{
 		/* We are trying to be smart here and conserve some vertical space.
 		   So we will skip tags not set, and try to show them in two parallel
-		   columns if they are short, which is by far the most common case. */
+		   columns if they are int, which is by far the most common case. */
 		int linelimit;
 		int climit[2];
